feat: pass dateFilter and setter to ShowFilters
