+++ frontend/src/components/pages/search/components/FilterBar/SubFilterField.tsx
@@ -1,8 +1,5 @@
-import { EVENT_ID } from 'modules/filters/constant';
 import { DateFilter, FilterState, Option } from 'modules/filters/interface';
 import React, { Fragment } from 'react';
-import { useIntl } from 'react-intl';
-import InputDateWithMagnifier from '../InputDateWithMagnifier';
 import ShowFilters from './ShowFilters';
 
 interface Props {
@@ -24,38 +21,10 @@ const SubFilterField: React.FC<Props> = ({
     return null;
   }
 
-  const intl = useIntl();
-
   const entriesFilters = Object.entries(filters);
 
   if (entriesFilters.length === 0) {
     return null;
-  }
-
-  if (String(entriesFilters[0][0]) === EVENT_ID) {
-    return (
-      <>
-        <div className="flex flex-col mt-4 desktop:mt-0 desktop:ml-5">
-          <div className="font-bold mb-2 text-lg">Agenda</div>
-          <div className="flex flex-row justify-between desktop:mt-0">
-            <InputDateWithMagnifier
-              value={dateFilter.beginDate}
-              onChange={event => {
-                setDateFilter({ beginDate: event.target.value, endDate: dateFilter.endDate });
-              }}
-              placeholder={intl.formatMessage({ id: 'search.beginDateFilter' })}
-            />
-            <InputDateWithMagnifier
-              value={dateFilter.endDate}
-              onChange={event => {
-                setDateFilter({ beginDate: dateFilter.beginDate, endDate: event.target.value });
-              }}
-              placeholder={intl.formatMessage({ id: 'search.endDateFilter' })}
-            />
-          </div>
-        </div>
-      </>
-    );
   }
 
   // Display filter items in a row
@@ -66,7 +35,12 @@ const SubFilterField: React.FC<Props> = ({
           <Fragment key={index}>
             {content.map(filter => (
               <div className="my-1" key={filter.id}>
-                <ShowFilters item={filter} setFilterSelectedOptions={setFilterSelectedOptions} />
+                <ShowFilters
+                  item={filter}
+                  setFilterSelectedOptions={setFilterSelectedOptions}
+                  dateFilter={dateFilter}
+                  setDateFilter={setDateFilter}
+                />
               </div>
             ))}
           </Fragment>
@@ -86,6 +60,8 @@ const SubFilterField: React.FC<Props> = ({
               key={filter.id}
               item={filter}
               setFilterSelectedOptions={setFilterSelectedOptions}
+              dateFilter={dateFilter}
+              setDateFilter={setDateFilter}
             />
           ))}
         </div>
